requirements.txt
pygame~=2.0.0
numpy==1.19.3
pymongo~=3.11.1
python-dotenv~=0.15.0